moved emails resource into an emails folder
